LPS-61290 move entries from layouts admin to mdr
